requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==4.2.11
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
executing==2.0.1
importlib_metadata==7.1.0
ipython==8.18.1
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
zipp==3.18.1
